Refactor dynamodb AQL-ish implementation
* Move the dynamodb sqlish parser to a instance from module level
* The class method of `parse` is moved to instance method.
* The instance of Statement class would be created without parameter.
* Remove dependency to the aws-dynamodb.js at all.
* Remove unused scripts
* Normalize module's path to be required
* Write jsdoc comments
