+++ bin/aws-dynamodb-query.js
@@ -1,7 +1,7 @@
 #!/bin/env node
 "use strict";
-var dynamodb = require("../lib/aws-dynamodb");
-dynamodb.connect();
+const awscli = require("../lib/awscli.js");
+const DynamoDBQueryStatement = require("../lib/dynamodb-query-statement.js");
 
 var Statement = require("../lib/dynamodb-statement.js");
 var ResultSet = require("../lib/dynamodb-result-set");
@@ -121,7 +121,9 @@ try {
         }
     }
 
-    var statement = dynamodb.QueryStatement(param);
+    const statement = new DynamoDBQueryStatement(param);
+    awscli.connect();
+    statement.dynamodb = awscli.getService("DynamoDB");
     var placeholderValues = {};
     if(getopt.options["placeholder-values"]) {
         placeholderValues = JSON.parse(getopt.options["placeholder-values"]);
