+++ lib/dynamodb-data-models.js
@@ -1,5 +1,5 @@
 "use strict";
-var DynamoDbDataType = require("./dynamodb-data-type");
+var DynamoDbDataType = require("./dynamodb-data-type.js");
 function DynamoDbDataModels() {}
 DynamoDbDataModels.obj2map = function(value) {
     if(value == "null") {
