+++ lib/dynamodb-delete-item-statement.js
@@ -1,6 +1,5 @@
 "use strict";
-var DynamoDbSqlishParser = require("../lib/dynamodb-sqlish-parser.js");
-var Statement = require('../lib/dynamodb-statement.js');
+var Statement = require('./dynamodb-statement.js');
 
 /**
  * SQL-ish DeleteItem statement class for AWS DynamoDB.
@@ -13,23 +12,27 @@ var Statement = require('../lib/dynamodb-statement.js');
  */
 function DynamoDbDeleteItemStatement(opt) {
     Statement.apply(this, Array.from(arguments));
+    this.key = {};
+    if(!opt) {
+        return;
+    }
     if(typeof(opt) === "string") {
-        opt = DynamoDbDeleteItemStatement.parse(opt);
+        opt = this.parse(opt);
     }
     if(!("TableName" in opt)) {
         throw new Error("TableName required");
     }
     this.setTableName(opt.TableName);
     if("Key" in opt) {
-        this.key = DynamoDbSqlishParser.parseItemListToMap(opt.Key);
+        this.key = this._parser.parseItemListToMap(opt.Key);
     }
 }
 
 DynamoDbDeleteItemStatement.prototype = new Statement();
 
-DynamoDbDeleteItemStatement.parse = function(sqlish) {
+DynamoDbDeleteItemStatement.prototype.parse = function(sqlish) {
     var opt = {};
-    var st = DynamoDbSqlishParser.parseDeleteItem(sqlish);
+    var st = this._parser.parseDeleteItem(sqlish);
 
     var tableName = st.getTerm("table-name");
     if(!tableName.match) {
@@ -77,4 +80,3 @@ DynamoDbDeleteItemStatement.prototype.getParameter = function(args) {
     return opt;
 };
 module.exports = DynamoDbDeleteItemStatement;
-
