+++ lib/dynamodb-put-item-statement.js
@@ -1,7 +1,6 @@
 "use strict";
-var DynamoDbSqlishParser = require("../lib/dynamodb-sqlish-parser.js");
-var Statement = require('../lib/dynamodb-statement.js');
-const DynamoDbDataModels = require("../lib/dynamodb-data-models.js");
+var Statement = require('./dynamodb-statement.js');
+const DynamoDbDataModels = require("./dynamodb-data-models.js");
 
 /**
  * SQL-ish PutItem statement class for AWS DynamoDB.
@@ -15,8 +14,13 @@ const DynamoDbDataModels = require("../lib/dynamodb-data-models.js");
 function DynamoDbPutItemStatement(opt) {
     Statement.apply(this, Array.from(arguments));
     this._itemAttr = null;
+    this.item = {};
+
+    if(!opt) {
+        return;
+    }
     if(typeof(opt) === "string") {
-        opt = DynamoDbPutItemStatement.parse(opt);
+        opt = this.parse(opt);
         this._itemAttr = {
             names: opt._attrNames,
             values: opt._attrNames
@@ -29,7 +33,7 @@ function DynamoDbPutItemStatement(opt) {
         throw new Error("Item required");
     }
     this.setTableName(opt.TableName);
-    this.item = DynamoDbSqlishParser.parseItemListToMap(opt.Item);
+    this.item = this._parser.parseItemListToMap(opt.Item);
     if("ConditionExpression" in opt) {
         this.conditionExpression =
             this.parseConditionExpression( opt.ConditionExpression );
@@ -39,9 +43,9 @@ function DynamoDbPutItemStatement(opt) {
 
 DynamoDbPutItemStatement.prototype = new Statement();
 
-DynamoDbPutItemStatement.parse = function(sqlish) {
+DynamoDbPutItemStatement.prototype.parse = function(sqlish) {
     var opt = {};
-    var st = DynamoDbSqlishParser.parsePutItem(sqlish);
+    var st = this._parser.parsePutItem(sqlish);
 
     var tableName = st.getTerm("table-name");
     if(!tableName.match) {
