+++ lib/dynamodb-query-statement.js
@@ -1,7 +1,6 @@
 "use strict";
-var Statement = require('../lib/dynamodb-statement.js');
-var DynamoDbReadItemStatement = require('../lib/dynamodb-read-item-statement.js');
-var DynamoDbSqlishParser = require("../lib/dynamodb-sqlish-parser.js");
+const Statement = require('./dynamodb-statement.js');
+const DynamoDbReadItemStatement = require('./dynamodb-read-item-statement.js');
 
 /**
  * SQL-ish Query statement class for AWS DynamoDB.
@@ -24,8 +23,11 @@ function DynamoDbQueryStatement(opt) {
     this._param = null;
     this._lastEvaluatedKey = null;
 
+    if(!opt) {
+        return;
+    }
     if(typeof(opt) === "string") {
-        opt = DynamoDbQueryStatement.parse(opt);
+        opt = this.parse(opt);
     }
     if(!("TableName" in opt)) {
         throw new Error("TableName required");
@@ -75,17 +77,22 @@ DynamoDbQueryStatement.prototype = new DynamoDbReadItemStatement();
 DynamoDbQueryStatement.prototype.run = function(args, callback) {
     this._callback = callback || ( (/*err, data*/) => {/*none*/} );
     this._param = this.getParameter(args);
-    Statement.assertAllParamSpecified(this._param);
-    this.dynamodb.query(this._param, (err, data) => { this.callback(err, data); });
+    this._query();
 };
 
-DynamoDbQueryStatement.prototype.callback = function(err, data) {
-    if(data != null && "LastEvaluatedKey" in data) {
-        this._lastEvaluatedKey = data.LastEvaluatedKey;
-    } else {
-        this._lastEvaluatedKey = null;
+DynamoDbQueryStatement.prototype._query = function() {
+    if(!this.dynamodb) {
+        throw new Error("Could not run a statement. The API is not connected.");
     }
-    this._callback(err, data);
+    Statement.assertAllParamSpecified(this._param);
+    this.dynamodb.query(this._param, (err, data) => {
+        if(data != null && "LastEvaluatedKey" in data) {
+            this._lastEvaluatedKey = data.LastEvaluatedKey;
+        } else {
+            this._lastEvaluatedKey = null;
+        }
+        this._callback(err, data);
+    });
 };
 
 /**
@@ -99,31 +106,40 @@ DynamoDbQueryStatement.prototype.callback = function(err, data) {
 DynamoDbQueryStatement.prototype.next = function() {
     if(this._lastEvaluatedKey) {
         this._param.ExclusiveStartKey = this._lastEvaluatedKey;
-        Statement.assertAllParamSpecified(this._param);
-        this.dynamodb.query(this._param, (err, data) => { this.callback(err, data); });
+        this._query();
     } else if(this._callback) {
         this._callback(null, null);
     }
 };
 
+/**
+ * Clear the previous query setup.
+ * Do not invoke next method after invoking this method.
+ * @returns {undefined}
+ */
 DynamoDbQueryStatement.prototype.reset = function() {
     this._callback = null;
     this._param = null;
     this._lastEvaluatedKey = null;
 };
 
-DynamoDbQueryStatement.parse = function(sqlish) {
-    var opt = {};
-    var st = DynamoDbSqlishParser.parseQuery(sqlish);
+/**
+ * Parse the SQL-ish statement.
+ * @param {string} sqlish SQL-ish statement
+ * @returns {object} A parameter for DynamoDB query API.
+ */
+DynamoDbQueryStatement.prototype.parse = function(sqlish) {
+    const opt = {};
+    const st = this._parser.parseQuery(sqlish);
 
-    var fromClause = st.getTerm("from-clause");
+    const fromClause = st.getTerm("from-clause");
     if(!fromClause.match) {
         throw new Error("the from-clause not found");
     } else {
         opt.TableName = fromClause.getWordsList("table-name")[0].join("");
     }
 
-    var whereClause = st.getTerm("where-key-clause");
+    const whereClause = st.getTerm("where-key-clause");
     if(!whereClause.match) {
         throw new Error("the where clause not found");
     } else {
@@ -131,25 +147,30 @@ DynamoDbQueryStatement.parse = function(sqlish) {
             whereClause.getWordsList("condition-expression")[0].join(" ");
     }
 
-    var selectClause = st.getTerm("select-clause");
+    const selectClause = st.getTerm("select-clause");
     if(selectClause.match) {
         opt.ProjectionExpression =
             selectClause.getWordsList("key-list")[0].join("");
     }
-    var filterClause = st.getTerm("filter-clause");
+    const filterClause = st.getTerm("filter-clause");
     if(filterClause.match) {
         opt.FilterExpression =
             filterClause.getWordsList("condition-expression")[0].join(" ");
     }
-    var limitClause = st.getTerm("limit-clause");
+    const limitClause = st.getTerm("limit-clause");
     if(limitClause.match) {
         opt.Limit = limitClause.getWordsList("limit-count")[0].join(" ");
     }
     return opt;
 };
 
+/**
+ * Get parameter to invoke the query API.
+ * @param {object} args key-values for the expression to run the query.
+ * @returns {object} A parameter for DynamoDB query API.
+ */
 DynamoDbQueryStatement.prototype.getParameter = function(args) {
-    var opt = DynamoDbReadItemStatement.prototype.getParameter.call(this, args);
+    const opt = DynamoDbReadItemStatement.prototype.getParameter.call(this, args);
     if(this.keyConditionExpression) {
         opt.KeyConditionExpression = this.keyConditionExpression;
     }
@@ -168,10 +189,20 @@ DynamoDbQueryStatement.prototype.getParameter = function(args) {
     return opt;
 };
 
+/**
+ * Set KeyConditionExpression for query.
+ * @param {string} keyConditionExpr expression
+ * @returns {undefined}
+ */
 DynamoDbQueryStatement.prototype.setKeyConditionExpression = function(keyConditionExpr) {
     this.keyConditionExpression = this.parseConditionExpression( keyConditionExpr );
 };
 
+/**
+ * Set FilterExpression for query.
+ * @param {string} filterExpr expression
+ * @returns {undefined}
+ */
 DynamoDbQueryStatement.prototype.setFilterExpression = function(filterExpr) {
     this.filterExpression = this.parseConditionExpression( filterExpr );
 };
