+++ lib/dynamodb-scan-statement.js
@@ -1,7 +1,6 @@
 "use strict";
-var Statement = require('../lib/dynamodb-statement.js');
-var DynamoDbReadItemStatement = require('../lib/dynamodb-read-item-statement.js');
-var DynamoDbSqlishParser = require("../lib/dynamodb-sqlish-parser.js");
+const Statement = require('./dynamodb-statement.js');
+const DynamoDbReadItemStatement = require('./dynamodb-read-item-statement.js');
 
 /**
  * SQL-ish Scan statement class for AWS DynamoDB.
@@ -29,7 +28,7 @@ function DynamoDbScanStatement(opt) {
     }
 
     if(typeof(opt) === "string") {
-        opt = DynamoDbScanStatement.parse(opt);
+        opt = this.parse(opt);
     }
     if(!("TableName" in opt)) {
         throw new Error("TableName required");
@@ -75,17 +74,22 @@ DynamoDbScanStatement.prototype = new DynamoDbReadItemStatement();
 DynamoDbScanStatement.prototype.run = function(args, callback) {
     this._callback = callback || ( (/*err, data*/) => {/*none*/} );
     this._param = this.getParameter(args);
-    Statement.assertAllParamSpecified(this._param);
-    this.dynamodb.scan(this._param, (err, data) => { this.callback(err, data); });
+    this._scan();
 };
 
-DynamoDbScanStatement.prototype.callback = function(err, data) {
-    if(data != null && "LastEvaluatedKey" in data) {
-        this._lastEvaluatedKey = data.LastEvaluatedKey;
-    } else {
-        this._lastEvaluatedKey = null;
+DynamoDbScanStatement.prototype._scan = function() {
+    if(!this.dynamodb) {
+        throw new Error("Could not run a statement. The API is not connected.");
     }
-    this._callback(err, data);
+    Statement.assertAllParamSpecified(this._param);
+    this.dynamodb.scan(this._param, (err, data) => {
+        if(data != null && "LastEvaluatedKey" in data) {
+            this._lastEvaluatedKey = data.LastEvaluatedKey;
+        } else {
+            this._lastEvaluatedKey = null;
+        }
+        this._callback(err, data);
+    });
 };
 
 /**
@@ -99,51 +103,65 @@ DynamoDbScanStatement.prototype.callback = function(err, data) {
 DynamoDbScanStatement.prototype.next = function() {
     if(this._lastEvaluatedKey) {
         this._param.ExclusiveStartKey = this._lastEvaluatedKey;
-        Statement.assertAllParamSpecified(this._param);
-        this.dynamodb.scan(this._param, (err, data) => { this.callback(err, data); });
+        this._scan();
     } else if(this._callback) {
         this._callback(null, null);
     }
 };
 
+/**
+ * Clear the previous scan setup.
+ * Do not invoke next method after invoking this method.
+ * @returns {undefined}
+ */
 DynamoDbScanStatement.prototype.reset = function() {
     this._callback = null;
     this._param = null;
     this._lastEvaluatedKey = null;
 };
 
-DynamoDbScanStatement.parse = function(sqlish) {
-    var opt = {};
-    var st = DynamoDbSqlishParser.parseScan(sqlish);
+/**
+ * Parse the SQL-ish statement.
+ * @param {string} sqlish SQL-ish statement
+ * @returns {object} A parameter for DynamoDB scan API.
+ */
+DynamoDbScanStatement.prototype.parse = function(sqlish) {
+    const opt = {};
+    const st = this._parser.parseScan(sqlish);
 
-    var fromClause = st.getTerm("from-clause");
+    const fromClause = st.getTerm("from-clause");
     if(!fromClause.match) {
         throw new Error("the from-clause not found");
     } else {
         opt.TableName = fromClause.getWordsList("table-name")[0].join("");
     }
 
-    var whereClause = st.getTerm("where-filter-clause");
+    const whereClause = st.getTerm("where-filter-clause");
     if(whereClause.match) {
         opt.FilterExpression =
             whereClause.getWordsList("condition-expression")[0].join(" ");
     }
 
-    var selectClause = st.getTerm("select-clause");
+    const selectClause = st.getTerm("select-clause");
     if(selectClause.match) {
         opt.ProjectionExpression =
             selectClause.getWordsList("key-list")[0].join("");
     }
 
-    var limitClause = st.getTerm("limit-clause");
+    const limitClause = st.getTerm("limit-clause");
     if(limitClause.match) {
         opt.Limit = limitClause.getWordsList("limit-count")[0].join(" ");
     }
     return opt;
 };
 
+/**
+ * Get parameter to invoke the scan API.
+ * @param {object} args key-values for the expression to run the scan.
+ * @returns {object} A parameter for DynamoDB scan API.
+ */
 DynamoDbScanStatement.prototype.getParameter = function(args) {
-    var opt = DynamoDbReadItemStatement.prototype.getParameter.call(this, args);
+    const opt = DynamoDbReadItemStatement.prototype.getParameter.call(this, args);
     if(this.limit) {
         opt.Limit = this.limit;
     }
@@ -159,6 +177,11 @@ DynamoDbScanStatement.prototype.getParameter = function(args) {
     return opt;
 };
 
+/**
+ * Set FilterExpression for scan.
+ * @param {string} filterExpr expression
+ * @returns {undefined}
+ */
 DynamoDbScanStatement.prototype.setFilterExpression = function(filterExpr) {
     this.filterExpression = this.parseConditionExpression( filterExpr );
 };
