+++ test/dynamodb-scan-statement.test.js
@@ -2,9 +2,16 @@
 const assert = require("chai").assert;
 const ScanStatement = require("../lib/dynamodb-scan-statement.js");
 describe("DynamoDbScanStatement", () => {
+    describe("constructor", () => {
+        it("should not be thrown without parameter", () => {
+            assert.doesNotThrow(()=>{
+                new ScanStatement();
+            });
+        })
+    });
     describe("parse", () => {
         it("can parse SQL-ish that all clause is included", function() {
-            var param = ScanStatement.parse(
+            var param = (new ScanStatement()).parse(
                 "SELECT A, B, C \r\n" +
                 "FROM TBL \n" +
                 "WHERE F=FilterValue Limit\n" +
@@ -15,7 +22,7 @@ describe("DynamoDbScanStatement", () => {
             assert.equal("10", param.Limit);
         });
         it("can parse SQL-ish that does not have optional clause", function() {
-            var param = ScanStatement.parse(
+            var param = (new ScanStatement()).parse(
                     " FROM TBL");
             assert.equal(false, "ProjectionExpression" in param);
             assert.equal("TBL", param.TableName);
@@ -23,7 +30,7 @@ describe("DynamoDbScanStatement", () => {
             assert.equal(false, "Limit" in param);
         });
         it("can parse SQL-ish that has various pattern", function() {
-            var param = ScanStatement.parse(
+            var param = (new ScanStatement()).parse(
                     "  SELECT A,B,C FROM TBL Limit 10   ");
             assert.equal("A,B,C", param.ProjectionExpression);
             assert.equal("TBL", param.TableName);
@@ -31,7 +38,7 @@ describe("DynamoDbScanStatement", () => {
             assert.equal("10", param.Limit);
         });
         it("can parse SQL-ish that has various pattern", function() {
-            var param = ScanStatement.parse(
+            var param = (new ScanStatement()).parse(
                     "FROM stars WHERE mainStar='SUN' AND orbitOrder BETWEEN 1 AND 9\n" +
                     "AND mass<1");
             assert.equal(false, "ProjectionExpression" in param);
